+++ src/tools/index.ts
@@ -1,6 +1,7 @@
 export * from "./_shared";
 export * from "./arp";
 export * from "./better-effects-panel";
+export * from "./better-inventory";
 export * from "./better-trade";
 export * from "./condition-manager";
 export * from "./giveth";